Make Reaction type assertions wordsize-agnostic (#1552)
On 32-bit Julia, `ones(Int, ...)` stoichiometries and `2//3` literals
produce Int32/Rational{Int32}, so `rx1 isa Reaction{SymbolicT,Int64}`
and `rx4 isa Reaction{SymbolicT,Rational{Int64}}` fail on the x86 CI
lane. Assert `Int`/`Rational{Int}` instead.

Verified on Julia 1.13.0 x86: rx1 is Reaction{SymbolicT,Int32} and the
updated assertions pass for all rate types.



Agent-Harness: Devin CLI 3000.10.21
Agent-Model: SWE-2 Max
Agent-Session: local Devin CLI session (no shareable URL); workspace /home/crackauc/sandbox/tmp_20260911_110613_57839

+++ test/reactionsystem_core/reaction.jl
@@ -30,10 +30,10 @@ let
         rx6 = Reaction(rate, [X], [x], [n1], [1])
 
         # Check `Reaction` types.
-        @test rx1 isa Reaction{SymbolicT,Int64}
+        @test rx1 isa Reaction{SymbolicT,Int}
         @test rx2 isa Reaction{SymbolicT,Float64}
         @test rx3 isa Reaction{SymbolicT,Any}
-        @test rx4 isa Reaction{SymbolicT,Rational{Int64}}
+        @test rx4 isa Reaction{SymbolicT,Rational{Int}}
         @test rx5 isa Reaction{SymbolicT,Any}
         @test rx6 isa Reaction{SymbolicT,Any}
